Split dbs_test.rb into focused test files for better CI parallelization
The original dbs_test.rb contained 40 tests covering multiple areas of
database rake tasks. By splitting into focused files, CI can distribute
tests more evenly across parallel workers:

- dbs_test.rb: Core db:create/db:drop tests (9 tests)
- dbs_schema_test.rb: Schema dump/load/cache and migration tests (19 tests)
- dbs_setup_test.rb: Setup, prepare, seed, and truncate tests (12 tests)
- dbs_postgresql_test.rb: PostgreSQL-specific tests (3 tests, from previous commit)

This allows the CI scheduler to parallelize what was previously a single
20+ minute bottleneck across multiple workers.
